backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.1
Flask-Bcrypt==1.0.1
Flask-Limiter==3.7.0
Flask-Caching==2.3.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
cloudinary==1.40.0
requests==2.32.3
gunicorn==22.0.0
resend==2.4.0
imageio-ffmpeg==0.5.1
